refactor: Content 제거, 에러 발생 시 navigate 제거
