+++ src/App.tsx
@@ -1,11 +1,10 @@
 import { Routes, Route } from 'react-router-dom';
-import { ThemeProvider } from 'styled-components';
+import styled, { ThemeProvider } from 'styled-components';
 
 import GlobalStyle from './styles/GlobalStyle';
 import theme from './styles/theme';
 
-import Header from './components/Layout/Header/Header';
-import Content from './components/Layout/Content/Content';
+import Header from './components/Header/Header';
 import MainPage from './pages/MainPage/MainPage';
 import ProductPage from './pages/ProductPage/ProductPage';
 import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
@@ -16,16 +15,25 @@ function App() {
       <GlobalStyle />
       <ThemeProvider theme={theme}>
         <Header />
-        <Content>
+        <StyledContent>
           <Routes>
             <Route path="/" element={<MainPage />} />
             <Route path="/product/:id" element={<ProductPage />} />
             <Route path="*" element={<NotFoundPage />} />
           </Routes>
-        </Content>
+        </StyledContent>
       </ThemeProvider>
     </>
   );
 }
+
+const StyledContent = styled.div`
+  min-height: 200px;
+  margin: 60px 0;
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+`;
 
 export default App;
