--- src/components/Layout/Content/Content.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-const StyledContent = styled.div`
-  min-height: 200px;
-  margin: 60px 0;
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-`;
-
-export default StyledContent;
