+++ src/configs/api.ts
@@ -1,3 +1,3 @@
-const SERVER_URL = 'https://json-server-shopping-cart.herokuapp.com';
+const SERVER_URL = 'https://json-server-shopping-cart.herokuapp.co';
 
 export default SERVER_URL;
