+++ src/pages/MainPage/MainPage.tsx
@@ -1,19 +1,13 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
 import ProductCardGrid from '../../components/ProductCardGrid/ProductCardGrid';
 import Spinner from '../../components/Spinner/Spinner';
 import useProductList from './useProductList';
 
 function MainPage() {
-  const navigate = useNavigate();
   const { isLoading, productList, error } = useProductList();
 
-  useEffect(() => {
-    if (error) {
-      alert(error);
-      navigate('/');
-    }
-  }, [error, navigate]);
+  if (error) {
+    alert(error);
+  }
 
   if (isLoading) return <Spinner />;
